robot-mount/requirements.txt
numpy
scipy
pyserial
sounddevice
SoundFile
tqdm